refactor(frontend): extract document and workspace api clients
