+++ frontend/src/modules/documents/infrastructure/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./documents.api";
