+++ frontend/src/modules/workspaces/infrastructure/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./workspaces.api";
